Gestion des erreurs

+++ src/routes/deletePokemon.js
@@ -3,6 +3,10 @@ const { Pokemon } = require('../db/sequelize')
 module.exports = (app) => {
     app.delete('/api/pokemons/:id', (req, res) => {
         Pokemon.findByPk(req.params.id).then(pokemon => {
+            if (pokemon === null) {
+                const message = 'Le pokémon demandé n\'existe pas. Réessayer avec un autre identifiant'
+                return res.status(404).json(message)
+            }
             const pokemonDeleted = pokemon;
             Pokemon.destroy({
                 where: { id: pokemon.id }
@@ -12,5 +16,9 @@ module.exports = (app) => {
                     res.json({ message, data: pokemonDeleted })
                 })
         })
+            .catch(error => {
+                const message = "Le pokémon n'a pas pu être modifié. Réessayez dans quelques instants."
+                res.status(500).json({ message, data: error })
+            })
     })
 }
